Refactor show_wrong_file_warning
